:construction: [WIP]dbMessagesのtagsに各々のtime値を保存処理

+++ components/Message.vue
@@ -37,7 +37,7 @@
 
 <script>
 	import MessageModel from '../models/Message';
-	import { dbMessages } from '../plugins/firebase';
+	import firebase, { dbMessages, dbTags } from '../plugins/firebase';
 	import Button from './Button';
 
 	export default {
@@ -71,7 +71,35 @@
 				try {
 					const id = this.i;
 					const docIds = await MessageModel.clear();
-					await dbMessages.doc(docIds[id]).delete();
+					const Data = await dbMessages.doc(docIds[id]).get();
+
+					Data.data().tags.map(async Element => {
+						Element = JSON.parse(JSON.stringify(Element));
+
+						const uid = firebase.auth().currentUser.uid;
+						const sameTagText = await dbTags.where('uid', '==', uid).where('text', '==', Element.text).get();
+						sameTagText.docs.map(async doc => {
+							// クリックした箇所の元々のdataとid
+							let sameTagTime = doc.data().time;
+							sameTagTime -= Element.time;
+
+							if (sameTagTime <= 0) {
+								await dbTags.doc(doc.id).delete();
+							} else {
+								await dbTags.doc(doc.id).update({
+									time: sameTagTime
+								});
+							}
+							return doc.id, doc.data();
+						});
+					});
+					if (docIds[id]) {
+						await dbMessages.doc(docIds[id]).delete();
+					}
+
+					// ここで強制リロード
+					this.$nextTick();
+
 					this.$emit('clear');
 				} catch (error) {
 					console.error(error);
